networking: Add multi-client server
Added the modifications mentioned in the pull-request review by Ghennadi Procopciuc
Fixed a bug where entering a blank line caused the client to hang.
Fixed a bug where long input caused the client and server to go out of sync. 'ACK for ' portion of the server response was removed.

Issue: ALB-751
Signed-off-by: Teodor Marina <teodor.marina@nxp.com>
Signed-off-by: Nica Dan <dan.nica@nxp.com>