clustering: Ensure the C/C++ MPI wrapper calls the correct compiler
If the MPI package was cross-compiled, the wrappers mpicc
and mpic++ will call the compiler that was used to build the
MPI package, using the flags used at buil time.

Therefore, the Makefile of the clustering demo should call the
C++ wrapper using the -cxx=g++ flag, to ensure the proper compiler
is used.

Issue: ALB-1446
Signed-off-by: Nica Dan <dan.nica@nxp.com>